Fixed Adding Questions
Fixed bug where you couldn't add questions with any answers that have a space in them.  Attached the edit question to the same thing that addQuestion uses.

Also added a search to see if a certain question is active based on it's ID, so if you search for an inactive question you will get an error message.